Install DuckDB Extensions Through HTTPS

The default core extension repository works through HTTP. There may be network limitations on directly downloading a file in this way (such as in an enterprise setting). To get around this, you can just add an 'S' in the URL to make it use HTTPS. This function does that automatically for you.

Usage

install_duckdb_extension(
  duckdb_connection = duck_con(),
  extension_name,
  duckdb_version,
  platform_name
)

Arguments

duckdb_connection

The DuckDB connection. Defaults the the in-memory database.

extension_name

Name of the extension to install. A list of extensions can be found in the DuckDB documentation.

duckdb_version

(Only needed for installing httpfs.) The version of DuckDB installed, don't include the 'v'. It can be found with packageVersion("duckdb").

platform_name

(Only needed for installing httpfs.) See the platforms for reference. This is checked against the list of platforms from the reference.

Details

More info about DuckDB extension installation.